requirements.txt
django==5.1.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-cors-headers==4.5.0
django-model-utils==5.0.0
django-filter==24.3
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
huey==2.5.1
drf-orjson-renderer==1.7.3